+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## v0.4.0
+
+Add warning when time series contain negative values and logarithmic scale is selected.
+
+## v0.3.0
+
+Added legend with time range specific color gradient, added switch to show / hide legend
+
 ## v0.2.0
 
 Added option to switch between linear and logarithmic scale